added method AddIngredient

+++ Ingredient.cs
@@ -9,10 +9,10 @@ namespace Cookbook
     class Ingredient
     {
         public string Name { get; set; }
-        public int Quantity { get; set; }
+        public double Quantity { get; set; }
         public string Unit { get; set; }
 
-        public Ingredient(string name, int quantity, string unit)
+        public Ingredient(string name, double quantity, string unit)
         {
             Name = name;
             Quantity = quantity <= 0 ? 0 : quantity;
